# Made bloom_filter type of index supporting arrays