--- a/dbms/src/Storages/MergeTree/MergeTreeIndexConditionBloomFilter.h
+++ b/dbms/src/Storages/MergeTree/MergeTreeIndexConditionBloomFilter.h
@@ -19,6 +19,7 @@ public:
             /// Atoms of a Boolean expression.
             FUNCTION_EQUALS,
             FUNCTION_NOT_EQUALS,
+            FUNCTION_HAS,
             FUNCTION_IN,
             FUNCTION_NOT_IN,
             FUNCTION_UNKNOWN, /// Can take any value.