--- a/dbms/tests/queries/0_stateless/00908_bloom_filter_index.reference
+++ b/dbms/tests/queries/0_stateless/00908_bloom_filter_index.reference
@@ -39,3 +39,4 @@
 6	some string
 13	abc
 		"rows_read": 3,
+1