--- a/dbms/tests/queries/0_stateless/00908_bloom_filter_index.sh
+++ b/dbms/tests/queries/0_stateless/00908_bloom_filter_index.sh
@@ -141,3 +141,13 @@ $CLICKHOUSE_CLIENT --query="SELECT * FROM bloom_filter_idx3 WHERE s IN ('some st
 $CLICKHOUSE_CLIENT --query="DROP TABLE bloom_filter_idx"
 $CLICKHOUSE_CLIENT --query="DROP TABLE bloom_filter_idx2"
 $CLICKHOUSE_CLIENT --query="DROP TABLE bloom_filter_idx3"
+
+$CLICKHOUSE_CLIENT --query="DROP TABLE IF EXISTS bloom_filter_idx_na;"
+$CLICKHOUSE_CLIENT -n --query="
+SET allow_experimental_data_skipping_indices = 1;
+CREATE TABLE bloom_filter_na
+(
+    na Array(Array(String)),
+    INDEX bf na TYPE bloom_filter(0.1) GRANULARITY 1
+) ENGINE = MergeTree()
+ORDER BY na" 2>&1 | grep -c 'DB::Exception: Unexpected type Array(Array(String)) of bloom filter index'