… dbms/tests/queries/0_stateless/00945_bloom_filter_index.reference