I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Category:
* New Feature

Short description:
Since current bloom_filter doesn't support array, so I made bloom_filter type of index supporting arrays.

### To use bloom_filter() for array, which I developed

1. Create a table with bloom_filter.

```SQL
SET allow_experimental_data_skipping_indices = 1;

CREATE TABLE default.test_bloomfilter
(
    `key` Array(String),
    `key_no_idx` Array(String),
    `id` String,
    `dt` DateTime,
    INDEX key_bf key TYPE bloom_filter(0.01) GRANULARITY 8192
)
ENGINE = MergeTree()
PARTITION BY toYYYYMMDD(dt)
ORDER BY id;
```

2. Insert data to the table.
```SQL
INSERT INTO test_bloomfilter VALUES (['k1', 'k2'], ['k1', 'k2'], 'id1', '2019-01-01 00:00:00'); 
INSERT INTO test_bloomfilter VALUES (['k2', 'k3'], ['k2', 'k3'], 'id2', '2019-01-02 00:00:00'); 
INSERT INTO test_bloomfilter VALUES (['k3', 'k4'], ['k3', 'k4'], 'id3', '2019-01-03 00:00:00');
```

3. Select data with bloom_filter.
```SQL
SELECT * from test_bloomfilter where has(key, 'k1');
```

### Search performance test
I did search performance test between with bloom_filter which I expanded and without bloom_filter. Search performance with bloom_filter which I expanded  outweighed.

```SQL
CREATE TABLE default.test_bloomfilter3 
(
  `key` Array(String), 
  `key_no_idx` Array(String), 
  value Array(String), 
  `id` String, 
  `dt` DateTime,  
  INDEX key_bf key TYPE bloom_filter(0.01) GRANULARITY 8192
) 
ENGINE = MergeTree() 
  PARTITION BY toYYYYMMDD(dt) 
  ORDER BY id
```

```SQL
INSERT INTO test_bloomfilter3 SELECT 
    [r, r + 1, r + 2, r + 3, r + 4, r + 5, r + 6, r + 7] AS key, 
    key AS key_no_idx, 
    key AS value, 
    r AS id, 
    toDateTime(p) AS dt
FROM 
(
    SELECT 
        number AS r, 
        r % 10000000 AS p
    FROM system.numbers
    LIMIT 10000000
)
```

**Without bloom_filter for array**
```SQL
WITH '10896' as k select value[indexOf(key, k)] from test_bloomfilter3 where has(key_no_idx, k)
```
8 rows in set. Elapsed: **0.148 sec**. 
Processed 10.00 million rows, 
1.35 GB (67.74 million rows/s., 9.15 GB/s.)

**With bloom_filter for array**
```SQL
WITH '10896' as k select value[indexOf(key, k)] from test_bloomfilter3 where has(key, k)
```
8 rows in set. Elapsed: **0.031 sec**. 
Processed 127.36 thousand rows, 
16.63 MB (4.12 million rows/s., 538.33 MB/s.)