@achimbab 2019-09-23
@zhang2014 
Thank you for your comments.
I'm felling that your insight for CH is great.
I fixed source codes.