@zhang2014 2019-09-23
👍 The code looks great.

BTW: Array(Array) currently does not appear to be supported. maybe we could consider using Array(bloom_filter_bits) to save array bloom filter?this way we  will support nesting of any number of Array types and the storage costs will always be the same. because:`(array_row_1_size + array_row_2_size + array_row_3_size) * per_row_need_bits_for_bloom_filter == array_row_1_size * per_row_need_bits_for_bloom_filter + array_row_2_size * per_row_need_bits_for_bloom_filter + array_row_3_size * per_row_need_bits_for_bloom_filter`

In addition, this method can easily add index support for other array functions in the future : )