@achimbab 2019-09-25
@zhang2014 
I'm appreciating for your code review.
It looks good that the idea extending this method to other skip indices.
**In the future**, it is possible to contribute the support of nested array type. :)