@zhang2014 2019-09-26
@alesapin @alexey-milovidov 
I'm not sure if this PR needs some other modifications, but it has almost solved all my doubts.