@achimbab 2019-10-04
@alesapin @alexey-milovidov 
We are using ClickHouse in 30 machines of cluster and now considering for expanding to 90 machines for the new features.
We need the bloomfilter for Array for the new features. 
Let me know what I should do for merging this PR.
Thank you.